Dropping GETX...
I will only focus on the StreamBuilder version, for now.
Maybe in the future, I can bring the GetX back to the project.

Reason why I'm dropping it: It has been consuming a lot of my time to maintain both State Managements. GetX is amazing, but now I will try to focus more on the Flutter StreamBuilder.